requirements.txt
# YOLO目标检测项目依赖库
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO实现
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.8.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 系统工具
tqdm>=4.64.0
pyyaml>=6.0
pathlib2>=2.3.0


# 安装命令：pip install -r requirements.txt